Add support for Naga v29
