+++ crates/ide/Cargo.toml
@@ -16,7 +16,8 @@ itertools.workspace = true
 line-index.workspace = true
 naga27 = { package = "naga", version = "27.0.3", features = ["wgsl-in"] }
 naga28 = { package = "naga", version = "28.0.0", features = ["wgsl-in"] }
-nagamain = { package = "naga", git = "https://github.com/gfx-rs/wgpu", rev = "2b8b5f7cdc1872a412371fc458b100de4d88068d", version = "28.0.0", features = [
+naga29 = { package = "naga", version = "29.0.1", features = ["wgsl-in"] }
+nagamain = { package = "naga", git = "https://github.com/gfx-rs/wgpu", rev = "ff438f3cadb8a25e0f36409566a6fb50f6707e02", features = [
 	"wgsl-in",
 ] }
 rowan.workspace = true
