+++ crates/ide/src/diagnostics.rs
@@ -118,141 +118,15 @@ trait NagaError: error::Error {
     fn spans(&self) -> Box<dyn Iterator<Item = (Option<Range<usize>>, String)> + '_>;
     fn location(&self) -> Option<Range<usize>>;
 }
+mod naga27;
+mod naga28;
+mod naga29;
+mod naga_main;
 
-struct Naga27;
-impl Naga for Naga27 {
-    type Module = naga27::Module;
-    type ParseError = naga27::front::wgsl::ParseError;
-    type ValidationError = naga27::WithSpan<naga27::valid::ValidationError>;
-
-    fn parse(source: &str) -> Result<Self::Module, Self::ParseError> {
-        naga27::front::wgsl::parse_str(source)
-    }
-
-    fn validate(module: &Self::Module) -> Result<(), Self::ValidationError> {
-        let flags = naga27::valid::ValidationFlags::all();
-        let capabilities = naga27::valid::Capabilities::all();
-        let mut validator = naga27::valid::Validator::new(flags, capabilities);
-        validator.validate(module).map(drop)
-    }
-}
-
-impl NagaError for naga27::front::wgsl::ParseError {
-    fn spans(&self) -> Box<dyn Iterator<Item = (Option<Range<usize>>, String)> + '_> {
-        Box::new(
-            self.labels()
-                .map(|(span, label)| (span.to_range(), label.to_owned())),
-        )
-    }
-
-    fn location(&self) -> Option<Range<usize>> {
-        let (span, _) = self.labels().next()?;
-        span.to_range()
-    }
-}
-
-impl NagaError for naga27::WithSpan<naga27::valid::ValidationError> {
-    fn spans(&self) -> Box<dyn Iterator<Item = (Option<Range<usize>>, String)> + '_> {
-        Box::new(
-            self.spans()
-                .map(move |(span, label)| (span.to_range(), label.clone())),
-        )
-    }
-
-    fn location(&self) -> Option<Range<usize>> {
-        self.spans().next().and_then(|(span, _)| span.to_range())
-    }
-}
-
-struct Naga28;
-impl Naga for Naga28 {
-    type Module = naga28::Module;
-    type ParseError = naga28::front::wgsl::ParseError;
-    type ValidationError = naga28::WithSpan<naga28::valid::ValidationError>;
-
-    fn parse(source: &str) -> Result<Self::Module, Self::ParseError> {
-        naga28::front::wgsl::parse_str(source)
-    }
-
-    fn validate(module: &Self::Module) -> Result<(), Self::ValidationError> {
-        let flags = naga28::valid::ValidationFlags::all();
-        let capabilities = naga28::valid::Capabilities::all();
-        let mut validator = naga28::valid::Validator::new(flags, capabilities);
-        validator.validate(module).map(drop)
-    }
-}
-
-impl NagaError for naga28::front::wgsl::ParseError {
-    fn spans(&self) -> Box<dyn Iterator<Item = (Option<Range<usize>>, String)> + '_> {
-        Box::new(
-            self.labels()
-                .map(|(span, label)| (span.to_range(), label.to_owned())),
-        )
-    }
-
-    fn location(&self) -> Option<Range<usize>> {
-        let (span, _) = self.labels().next()?;
-        span.to_range()
-    }
-}
-
-impl NagaError for naga28::WithSpan<naga28::valid::ValidationError> {
-    fn spans(&self) -> Box<dyn Iterator<Item = (Option<Range<usize>>, String)> + '_> {
-        Box::new(
-            self.spans()
-                .map(move |(span, label)| (span.to_range(), label.clone())),
-        )
-    }
-
-    fn location(&self) -> Option<Range<usize>> {
-        self.spans().next().and_then(|(span, _)| span.to_range())
-    }
-}
-
-struct NagaMain;
-impl Naga for NagaMain {
-    type Module = nagamain::Module;
-    type ParseError = nagamain::front::wgsl::ParseError;
-    type ValidationError = nagamain::WithSpan<nagamain::valid::ValidationError>;
-
-    fn parse(source: &str) -> Result<Self::Module, Self::ParseError> {
-        nagamain::front::wgsl::parse_str(source)
-    }
-
-    fn validate(module: &Self::Module) -> Result<(), Self::ValidationError> {
-        let flags = nagamain::valid::ValidationFlags::all();
-        let capabilities = nagamain::valid::Capabilities::all();
-        let mut validator = nagamain::valid::Validator::new(flags, capabilities);
-        validator.validate(module).map(drop)
-    }
-}
-
-impl NagaError for nagamain::front::wgsl::ParseError {
-    fn spans(&self) -> Box<dyn Iterator<Item = (Option<Range<usize>>, String)> + '_> {
-        Box::new(
-            self.labels()
-                .map(|(span, label)| (span.to_range(), label.to_owned())),
-        )
-    }
-
-    fn location(&self) -> Option<Range<usize>> {
-        let (span, _) = self.labels().next()?;
-        span.to_range()
-    }
-}
-
-impl NagaError for nagamain::WithSpan<nagamain::valid::ValidationError> {
-    fn spans(&self) -> Box<dyn Iterator<Item = (Option<Range<usize>>, String)> + '_> {
-        Box::new(
-            self.spans()
-                .map(move |(span, label)| (span.to_range(), label.clone())),
-        )
-    }
-
-    fn location(&self) -> Option<Range<usize>> {
-        self.spans().next().and_then(|(span, _)| span.to_range())
-    }
-}
+use naga_main::NagaMain;
+use naga27::Naga27;
+use naga28::Naga28;
+use naga29::Naga29;
 
 fn emit<Error: NagaError>(
     database: &dyn HirDatabase,
@@ -362,6 +236,9 @@ pub fn diagnostics(
             },
             NagaVersion::Naga28 => {
                 naga_diagnostics::<Naga28>(database, file_id, config, &mut diagnostics);
+            },
+            NagaVersion::Naga29 => {
+                naga_diagnostics::<Naga29>(database, file_id, config, &mut diagnostics);
             },
             NagaVersion::NagaMain => {
                 naga_diagnostics::<NagaMain>(database, file_id, config, &mut diagnostics);
