+++ crates/wgsl-analyzer/src/config.rs
@@ -113,8 +113,10 @@ pub enum NagaVersionConfig {
     #[serde(rename = "0.27")]
     Naga27,
     #[serde(rename = "0.28")]
-    #[default]
     Naga28,
+    #[default]
+    #[serde(rename = "0.29")]
+    Naga29,
     #[serde(rename = "main")]
     NagaMain,
 }
@@ -600,6 +602,7 @@ impl Config {
             naga_version: match self.diagnostics_nagaVersion() {
                 NagaVersionConfig::Naga27 => NagaVersion::Naga27,
                 NagaVersionConfig::Naga28 => NagaVersion::Naga28,
+                NagaVersionConfig::Naga29 => NagaVersion::Naga29,
                 NagaVersionConfig::NagaMain => NagaVersion::NagaMain,
             },
         }
